dom/security/test/csp/test_bug1579094.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test if Wildcard CSP supports ExternalProtocol</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta meta http-equiv="Content-security-policy" content="frame-src SomeExternalProto://*">
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
<script class="testbody" type="text/javascript">
    SimpleTest.waitForExplicitFinish();

    document.addEventListener("securitypolicyviolation",()=>{
        ok(false, "Error: ExternalProtocol Was blocked");
        SimpleTest.finish();
    });

    window.addEventListener("load", ()=>{
        ok(true, "Error: ExternalProtocol was passed");
        SimpleTest.finish();
    });
</script>

<iframe src="SomeExternalProto:foo@bar.com">


</body>
</html>
